A few additions to the web mailer list from Steve
Working on the perl version of Steve Jenkins post white list. The only
difference being sorting with Perl. Needs a few variable additions and
a code change to handle sorting with perl or not.
It still dumps the post white.cidr into this directory - a commented
line near the bottom of the script - so you don’t accidentally stuff it
in to your working directory. Change it if you use this script?
I like this sort better- It puts the IPv6 addresses where they should be